Rubocop rspec
